Generalize pytest ordering plugin for install and post test suite
